refactor: 유저관련 파일 수정
