+++ src/main/java/com/sparta/springtrello/domain/user/dto/LoginRequestDto.java
@@ -4,9 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
-@Setter
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
