+++ src/main/java/com/sparta/springtrello/domain/user/dto/PasswordUpdateRequestDto.java
@@ -4,11 +4,10 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import lombok.Getter;
-import lombok.Setter;
+
 
 @Getter
-@Setter
-public class UpdatePasswordRequestDto {
+public class PasswordUpdateRequestDto {
 
     @NotBlank(message = "현재 비밀번호는 필수 입력 값입니다.")
     private String currentPassword;
