+++ src/main/java/com/sparta/springtrello/domain/user/dto/ProfileUpdateRequestDto.java
@@ -2,12 +2,10 @@ package com.sparta.springtrello.domain.user.dto;
 
 import jakarta.validation.constraints.Size;
 import lombok.Getter;
-import lombok.Setter;
+
 
 @Getter
-@Setter
-public class UpdateProfileRequestDto {
-
+public class ProfileUpdateRequestDto {
     @Size(max = 10, message = "이름은 최대 10글자까지 입력 가능합니다.")
     private String nickname;
