+++ src/main/java/com/sparta/springtrello/domain/user/dto/SignupRequestDto.java
@@ -4,12 +4,9 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
-@Setter
 public class SignupRequestDto {
-
     @NotBlank(message = "유저네임은 필수 입력 값입니다.")
     @Size(min = 4, max = 10, message = "유저네임은 4자 이상, 10자 이하이어야 합니다.")
     @Pattern(regexp = "^[a-z0-9]+$", message = "유저네임은 소문자 알파벳과 숫자만 포함해야 합니다.")
