+++ src/main/java/com/sparta/springtrello/exception/custom/common/UploadException.java
@@ -1,7 +1,7 @@
 package com.sparta.springtrello.exception.custom.common;
 
 import com.sparta.springtrello.common.ResponseCodeEnum;
-import com.sparta.springtrello.exception.CustomException;
+import com.sparta.springtrello.exception.custom.CustomException;
 
 public class UploadException extends CustomException {
     public UploadException(ResponseCodeEnum responseCode) {
