+++ src/main/java/com/sparta/springtrello/exception/custom/user/PasswordException.java
@@ -1,7 +1,7 @@
 package com.sparta.springtrello.exception.custom.user;
 
 import com.sparta.springtrello.common.ResponseCodeEnum;
-import com.sparta.springtrello.exception.CustomException;
+import com.sparta.springtrello.exception.custom.CustomException;
 
 public class PasswordException extends CustomException {
     public PasswordException(ResponseCodeEnum responseCode) {
