+++ src/main/java/com/sparta/springtrello/exception/custom/user/UserException.java
@@ -1,7 +1,7 @@
 package com.sparta.springtrello.exception.custom.user;
 
 import com.sparta.springtrello.common.ResponseCodeEnum;
-import com.sparta.springtrello.exception.CustomException;
+import com.sparta.springtrello.exception.custom.CustomException;
 
 public class UserException extends CustomException {
     public UserException(ResponseCodeEnum responseCode) {
