+++ src/main/resources/application.yml
@@ -31,7 +31,7 @@ logging.level:
 #  org.springframework.web: debug
 
 jwt-secret-key: ${JWT-SECRET-KEY}
-admin-password: ${MANAGER-PASSWORD}
+manager-password: ${MANAGER-PASSWORD}
 
 cloud:
   aws:
